moved the function getting the timestamp string to platform
